crop detection/requirements.txt
flask==3.1.0
flask_cors==5.0.0
torch
torchvision
psutil==5.9.0
pandas==2.2.3
numpy==1.26.3
matplotlib==3.10.0
torchsummary
pillow
